Restructuring the provider
